+++ src/Set/Sequence.php
@@ -19,15 +19,53 @@ final class Sequence implements Implementation
      * @psalm-mutation-free
      *
      * @param Implementation<I> $set
-     * @param int<1, max> $size
      * @param int<1, max> $min
      */
     private function __construct(
         private Implementation $set,
         private Integers $sizes,
-        private int $size,
         private int $min,
     ) {
+    }
+
+    #[\Override]
+    public function __invoke(
+        Random $random,
+        \Closure $predicate,
+        int $size,
+    ): \Generator {
+        $shrinker = new Sequence\Shrinker;
+        $detonate = new Sequence\Detonate;
+        $min = $this->min;
+        $bounds = static fn(array $sequence): bool => \count($sequence) >= $min;
+        $predicate = static fn(array $sequence): bool => /** @var list<I> $sequence */ $bounds($sequence) && $predicate($sequence);
+        $immutable = ($this->set)($random, static fn() => true, 1)
+            ->current()
+            ?->immutable() ?? false;
+        $yielded = 0;
+
+        do {
+            foreach (($this->sizes)($random, static fn() => true, $size) as $nextSize) {
+                if ($yielded === $size) {
+                    return;
+                }
+
+                /** @psalm-suppress ArgumentTypeCoercion */
+                $values = $this->generate($nextSize->unwrap(), $random);
+                $value = Value::of($values)
+                    ->mutable(!$immutable)
+                    ->predicatedOn($predicate);
+                $yieldable = $value->map($detonate);
+
+                if (!$yieldable->acceptable()) {
+                    continue;
+                }
+
+                yield $yieldable->shrinkWith($shrinker);
+
+                ++$yielded;
+            }
+        } while ($yielded < $size);
     }
 
     /**
@@ -48,7 +86,6 @@ final class Sequence implements Implementation
         return new self(
             $set,
             $sizes,
-            100,
             $sizes->min(),
         );
     }
@@ -69,59 +106,6 @@ final class Sequence implements Implementation
     }
 
     /**
-     * @psalm-mutation-free
-     */
-    #[\Override]
-    public function take(int $size): self
-    {
-        return new self(
-            $this->set,
-            $this->sizes->take($size),
-            $size,
-            $this->min,
-        );
-    }
-
-    #[\Override]
-    public function values(Random $random, \Closure $predicate): \Generator
-    {
-        $shrinker = new Sequence\Shrinker;
-        $detonate = new Sequence\Detonate;
-        $min = $this->min;
-        $bounds = static fn(array $sequence): bool => \count($sequence) >= $min;
-        $predicate = static fn(array $sequence): bool => /** @var list<I> $sequence */ $bounds($sequence) && $predicate($sequence);
-        $immutable = $this
-            ->set
-            ->values($random, static fn() => true)
-            ->current()
-            ?->immutable() ?? false;
-        $yielded = 0;
-
-        do {
-            foreach ($this->sizes->values($random, static fn() => true) as $size) {
-                if ($yielded === $this->size) {
-                    return;
-                }
-
-                /** @psalm-suppress ArgumentTypeCoercion */
-                $values = $this->generate($size->unwrap(), $random);
-                $value = Value::of($values)
-                    ->mutable(!$immutable)
-                    ->predicatedOn($predicate);
-                $yieldable = $value->map($detonate);
-
-                if (!$yieldable->acceptable()) {
-                    continue;
-                }
-
-                yield $yieldable->shrinkWith($shrinker);
-
-                ++$yielded;
-            }
-        } while ($yielded < $this->size);
-    }
-
-    /**
      * @param 0|positive-int $size
      *
      * @return list<Value<I>>
@@ -132,9 +116,10 @@ final class Sequence implements Implementation
             return [];
         }
 
-        return \array_values(\iterator_to_array($this->set->take($size)->values(
+        return \array_values(\iterator_to_array(($this->set)(
             $rand,
             static fn() => true,
+            $size,
         )));
     }
 }